requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13
keyboard==0.13.5
numpy==1.24.3
soundfile==0.12.1
scipy==1.10.1
openai-whisper==20231117
python-dotenv==1.0.0
pyspellchecker==0.7.2
language-tool-python==2.7.1
spacy==3.7.2 